+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Upcoming
 
+### 🔄 Changed
+
+# [2.2.8](https://github.com/GetStream/stream-chat-swift/releases/tag/2.2.8)
+_August 07, 2020_
+
 ### ✅ Added
 - Membership for a channel  [#387](https://github.com/GetStream/stream-chat-swift/pull/387)